requirements.txt
allure-behave==2.11.1
allure-python-commons==2.11.1
async-generator==1.10
attrs==22.1.0
behave==1.2.6
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==21.3
parse==1.19.0
parse-type==0.6.0
pip==22.3
pluggy==1.0.0
pycparser==2.21
pyparsing==3.0.9
pyodbc==4.0.35
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
selenium==4.5.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.4
wsproto==1.2.0